requirements.txt
google-adk
google-genai==1.27.0
python-dotenv
requests
fastapi[standard]
pydantic
uvicorn
